Net reduction of about 150 warnings when compiling with clang 15.0.6
devif.h, mnxdefs.h:
Update MRegion to reflect that width/height are unsigned.
Update ...EventMask fields in XWINDOW section of DspInterfaceRec to match parameter
   type used in XSelectInput() to which they are passed.
Add padding to explicitly align colors field (short among longs)

xbitmaps.h:
Reorder fields in LISP_CURSOR struct to avoid requiring padding for alignment
   constraints, and adjust initialization to match.
Declare cursors static const.
Move Lisp_Icon from xbitmaps.h to xmkicon.c

xinit.c, xinitdefs.h
Replace bound() (xwinmandefs.h) with static inline unsigned ubound() in xinit.c
Change signature of X_init() to use unsigned values for width, height, depth.
Add casts where necessary to avoid implicit signedness conversions.
Add extern declarations for all globals defined in xinit.c

xwinman.c, xwinmandefs.h
Remove bound() from xwinmandefs.h and use static inline unsigned ubound() in xwinman.c
Change signature of lisp_Xconfigure() to use unsigned width and height parameters.
Make GravSize unsigned to match its usage context.
Add casts where necessary to avoid implicit signedness conversions.
Add missing break to correct switch case fall-through warning.

dspif.c
Change LispDisplayRequestedWidth/Height to unsigned to reflect context of use.

xlspwin.c
Eliminate declaration for unused XEvent report.
Move global variable gcv, used only locally, to local variable in the procedures where needed.
Change various variables used in width/height calculation to unsigned to match usage context.
Add casts where necessary to avoid implicit signedness conversions.

xmkicon.c
Move Lisp_Icon from xbitmaps.h to xmkicon.c.
Move global XImage IconImage to local in function where it is required.
Add cast where necessary to avoid implicit signedness conversions.

xrdopt.c
Remove incorrect casts for signedness causing implicit sign conversion warnings.
Add extern declarations for all globals defined in xrdopt.c

xscroll.c
Replace bound() (xwinmandefs.h) with static inline signed sbound() in xscroll.c

initdsp.c, initdspdefs.h
Change signedness of various display variables from int to unsigned.
Change signature of init_display2() to unsigned display_max.
Add extern declarations for some globals defined in initdsp.c

xbbt.c
Add explicit casts for type warnings from MRegion width/height change.
Add explicit casts for type warnings in arguments to XPutImage().

+++ src/xbbt.c
@@ -31,8 +31,8 @@ unsigned long clipping_Xbitblt(DspInterface dsp, DLword *dummy, int x, int y, in
 {
   int temp_x, temp_y, LowerRightX, LowerRightY;
 
-  LowerRightX = dsp->Visible.x + dsp->Visible.width - 1;
-  LowerRightY = dsp->Visible.y + dsp->Visible.height - 1;
+  LowerRightX = dsp->Visible.x + (int)dsp->Visible.width - 1;
+  LowerRightY = dsp->Visible.y + (int)dsp->Visible.height - 1;
 
   /* display region of interest lower right x, y pixel */
   temp_x = x + w - 1;
@@ -48,7 +48,7 @@ unsigned long clipping_Xbitblt(DspInterface dsp, DLword *dummy, int x, int y, in
       (temp_y <= LowerRightY)) {
     XLOCK;
     XPutImage(dsp->display_id, dsp->DisplayWindow, dsp->Copy_GC, &dsp->ScreenBitmap, x, y,
-              x - dsp->Visible.x, y - dsp->Visible.y, w, h);
+              x - dsp->Visible.x, y - dsp->Visible.y, (unsigned)w, (unsigned)h);
     XFlush(dsp->display_id);
     XUNLOCK(dsp);
     return (1);
@@ -75,7 +75,7 @@ unsigned long clipping_Xbitblt(DspInterface dsp, DLword *dummy, int x, int y, in
   if ((w > 0) && (h > 0)) {
     XLOCK;
     XPutImage(dsp->display_id, dsp->DisplayWindow, dsp->Copy_GC, &dsp->ScreenBitmap, x, y,
-              x - dsp->Visible.x, y - dsp->Visible.y, w, h);
+              x - dsp->Visible.x, y - dsp->Visible.y, (unsigned)w, (unsigned)h);
     XFlush(dsp->display_id);
     XUNLOCK(dsp);
   }
